Default theme ported to use a palette
Creating a theme is now a case of choosing a colour palette and the its shades. Each component is styled using only the palette colours.

Currently only the default theme has been ported and the current palette needs decent colours choosing. There may be the case where the palette size needs increasing from 5>7 or even 9 colours as components such as the StateBrowser uses many shades of the same colour.