Allow channel name from location query params
We cannot use the location anchor to prefill the channel with a direct
private channel to some user nick. As "#foobar" will previll the public
channel "#foobar" and not "foobar"

This patch allow the query parameter "?channel=foobar" to prefill the
channel input with a direct private channel and not a public one.

This will not break existing links cause we do not change the previous
url anchor way.