fix chathistory 005 token name
This is a bug introduced in https://github.com/kiwiirc/kiwiirc/pull/1244 ;
Kiwi was modified to use a draft prefix for both the capability name
and the 005 token, but only the capability name was modified on the spec side.
Since the latest Ergo release now publishes the correct token name, use that
name unconditionally.